pyproject.toml
[tool.poetry]
name = "konoha"
version = "1.0.0"
description = "A tiny sentence/word tokenizer for Japanese text written in Python"
authors = ["himkt <himkt@cookpad.com>"]

[tool.poetry.dependencies]
python = "^3.6"
janome = {version = "^0.3.10", optional = true}
natto-py = {version = "^0.9.0", optional = true}
kytea = {version = "^0.1.4", optional = true}
sentencepiece = {version = "^0.1.85", optional = true}
sudachipy = {version = "^0.4.2", optional = true}
boto3 = {version = "^1.11.0", optional = true}

[tool.poetry.dev-dependencies]
python-language-server = "^0.31.2"
pytest = "^5.3.2"
flake8 = "^3.7.9"
rope = "^0.14.0"
mypy = "^0.761"
black = "^19.10b0"
ipython = "^7.10.2"

[tool.poetry.extras]
janome = ["janome"]
mecab = ["natto-py"]
kytea = ["kytea"]
sudachi = ["sudachipy", "sudachidict-core"]
sentencepiece = ["sentencepiece"]
remote = ["boto3"]
all = ["janome", "natto-py", "kytea", "sudachipy", "sentencepiece", "boto3"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
